Add criteria for minMatch
This change doesn't affect walking in any way.. it's only used for
direct calls to RSW.GetMyPos and only if the param is set to True. If
it's less then the given minMatch then exit with Point(-1,-1)

The value for `RSW.minMatch` ranges from `-1.0` to `1.0`, where -1 is
opposite "match", and 1.0 is perfect match.